putting all 3 portions together for XSLT
hey guys, i copied and pasted all of our XMLs together into one document so we can start working on the XSLT for the webpage. So from now on, when you make changes to your portion of the code, please do it in the FULLplaytext.xml file in your part of the document -- i separated all of our portions with really spaced-out comments, so you should be able to find your portion pretty easily!